Refactored and added new asserts to DiscountsUtilTest.
